chore: fix problem with api generator
